requirements.txt
# Flask舆情分析系统依赖包
# 生成时间: 2025-01-07
# 项目: Flask舆情分析系统

# ===== 核心Web框架 =====
Flask==3.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
itsdangerous==2.1.2
MarkupSafe==2.1.3
click==8.1.7
aiohttp
# ===== 数据处理 =====
pandas==2.1.3
numpy==1.26.2

# ===== 网络请求 =====
requests==2.31.0
urllib3==2.1.0

# ===== 数据库 =====
PyMySQL==1.1.1
SQLAlchemy==2.0.23

# ===== 自然语言处理 =====
snownlp==0.12.3
jieba==0.42.1

# ===== 图表可视化 =====
pyecharts==2.0.7

# ===== 网页爬虫 =====
beautifulsoup4==4.12.2
lxml==5.1.0

# ===== 任务调度 =====
APScheduler==3.11.0

# ===== AI模型接口 =====
openai==1.82.0
zhipuai

# ===== 机器学习 =====
scikit-learn==1.3.2
wordcloud==1.9.2

# ===== 图像处理 =====
Pillow==10.1.0

# ===== 系统监控 =====
psutil==5.9.6

# ===== 数据验证 =====
pydantic==2.11.5

# ===== 异步处理 =====
aiohttp==3.9.1

# ===== 加密 =====
bcrypt==4.0.1
cryptography

# ===== 环境变量 =====
python-dotenv==1.0.1

# ===== 日期时间 =====
python-dateutil==2.8.2

# ===== JSON处理 =====
jsonpath==0.82.2

# ===== 文件处理 =====
openpyxl==3.1.2

# ===== 压缩（可选） =====
Flask-Compress==1.14

# ===== 开发工具（可选，注释掉以减少依赖） =====
# pytest==7.4.3
# black==23.11.0
# flake8==6.1.0

# ===== 说明 =====
# 安装方法:
# pip install -r requirements.txt
#
# 如果在中国大陆，建议使用国内镜像源:
# pip install -r requirements.txt -i https://pypi.tuna.tsinghua.edu.cn/simple/
#
# 或者:
# pip install -r requirements.txt -i https://mirrors.aliyun.com/pypi/simple/
